add editDistance

+++ src/main/java/com/numb3r3/common/io/FileUtils.java
@@ -272,6 +272,7 @@ public class FileUtils {
     public static boolean deleteFolderContents(File folder) {
         System.out.println("Deleting content of: " + folder.getAbsolutePath());
         File[] files = folder.listFiles();
+        if (files == null) return true;
         for (File file : files) {
             if (file.isFile()) {
                 if (!file.delete()) {
